Extract the ConfigProvider from the LedgerFactory and integrate it into the Runner
